+++ Frends.MicrosoftSQL.ExecuteProcedure/Frends.MicrosoftSQL.ExecuteProcedure/Definitions/Enums.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// SQL transaction isolation levels.
 /// </summary>
-public enum SqlTransactionIsolationLevel 
+public enum SqlTransactionIsolationLevel
 {
     /// <summary>
     /// A different isolation level than the one specified is being used, but the level cannot be determined.
@@ -29,7 +29,7 @@ public enum SqlTransactionIsolationLevel
     /// A dirty read is possible, meaning that no shared locks are issued and no exclusive locks are honored.
     /// </summary>
     ReadUncommitted = 256,
-    
+
     /// <summary>
     /// Locks are placed on all data that is used in a query, preventing other users from updating the data. 
     /// Prevents non-repeatable reads but phantom rows are still possible.
@@ -40,7 +40,7 @@ public enum SqlTransactionIsolationLevel
     /// A range lock is placed on the System.Data.DataSet, preventing other users from updating or inserting rows into the dataset until the transaction is complete.
     /// </summary>
     Serializable = 1048576,
-    
+
     /// <summary>
     /// Reduces blocking by storing a version of data that one application can read while another is modifying the same data. 
     /// Indicates that from one transaction you cannot see changes made in other transactions, even if you requery.