+++ Frends.MicrosoftSQL.ExecuteProcedure/Frends.MicrosoftSQL.ExecuteProcedure/Definitions/Input.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Frends.MicrosoftSQL.ExecuteProcedure.Definitions;
 
@@ -18,6 +19,7 @@ public class Input
 
     /// <summary>
     /// Name of stored procedure to execute.
+    /// Should only contain the name of the procedure without, for example, database name or semicolons.
     /// </summary>
     /// <example>ExampleProcedure</example>
     [DisplayFormat(DataFormatString = "Sql")]
@@ -69,5 +71,4 @@ public class ProcedureParameter
     /// <example>SqlDbTypes.Auto</example>
     [DefaultValue(SqlDataTypes.Auto)]
     public SqlDataTypes SqlDataType { get; set; }
-
 }