+++ Frends.MicrosoftSQL.ExecuteProcedure/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.1] - 2023-08-03
+### Changed
+- Documentation update to Input.Execute parameter.
+- Removed unnecessary runtime unloader.
+
 ## [1.0.0] - 2023-01-23
 ### Added
 - Initial implementation